Sandbox Az/runner_service/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
azure-identity==1.13.0
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.2.0
pydantic==2.5.0
psutil==5.9.5